# Adding Pandas MultiIndex cache