--- a/Common/Python/PandasData.cs
+++ b/Common/Python/PandasData.cs
@@ -36,7 +36,7 @@ namespace QuantConnect.Python
         private readonly int _levels;
         private readonly bool _isCustomData;
         private readonly Symbol _symbol;
-        private readonly Dictionary<string, Tuple<List<PyObject>, List<object>>> _series;
+        private readonly Dictionary<string, Tuple<List<DateTime>, List<object>>> _series;
 
         private readonly IEnumerable<MemberInfo> _members;
 
@@ -111,7 +111,7 @@ namespace QuantConnect.Python
                 columns.AddRange(keys);
             }
 
-            _series = columns.Distinct().ToDictionary(k => k, v => Tuple.Create(new List<PyObject>(), new List<object>()));
+            _series = columns.Distinct().ToDictionary(k => k, v => Tuple.Create(new List<DateTime>(), new List<object>()));
         }
 
         /// <summary>
@@ -123,11 +123,7 @@ namespace QuantConnect.Python
             foreach (var member in _members)
             {
                 var key = member.Name.ToLower();
-                PyObject endTime;
-                using (Py.GIL())
-                {
-                    endTime = (baseData as IBaseData).EndTime.ToPython();
-                }
+                var endTime = (baseData as IBaseData).EndTime;
                 AddToSeries(key, endTime, (member as FieldInfo)?.GetValue(baseData));
                 AddToSeries(key, endTime, (member as PropertyInfo)?.GetValue(baseData));
             }
@@ -135,11 +131,7 @@ namespace QuantConnect.Python
             var storage = (baseData as DynamicData)?.GetStorageDictionary();
             if (storage != null)
             {
-                PyObject endTime;
-                using (Py.GIL())
-                {
-                    endTime = (baseData as IBaseData).EndTime.ToPython();
-                }
+                var endTime = (baseData as IBaseData).EndTime;
                 var value = (baseData as IBaseData).Value;
                 AddToSeries("value", endTime, value);
 
@@ -167,11 +159,7 @@ namespace QuantConnect.Python
         {
             if (tradeBar != null)
             {
-                PyObject time;
-                using (Py.GIL())
-                {
-                    time = tradeBar.EndTime.ToPython();
-                }
+                var time = tradeBar.EndTime;
                 AddToSeries("open", time, tradeBar.Open);
                 AddToSeries("high", time, tradeBar.High);
                 AddToSeries("low", time, tradeBar.Low);
@@ -180,11 +168,7 @@ namespace QuantConnect.Python
             }
             if (quoteBar != null)
             {
-                PyObject time;
-                using (Py.GIL())
-                {
-                    time = quoteBar.EndTime.ToPython();
-                }
+                var time = quoteBar.EndTime;
                 if (tradeBar == null)
                 {
                     AddToSeries("open", time, quoteBar.Open);
@@ -215,11 +199,7 @@ namespace QuantConnect.Python
                 {
                     if (tick == null) continue;
 
-                    PyObject time;
-                    using (Py.GIL())
-                    {
-                        time = tick.EndTime.ToPython();
-                    }
+                    var time = tick.EndTime;
                     var column = tick.TickType == TickType.OpenInterest
                         ? "openinterest"
                         : "lastprice";
@@ -283,12 +263,13 @@ namespace QuantConnect.Python
                 var isFalse = x is bool && !(bool)x;
                 return x == null || isNaNOrZero || isNullOrWhiteSpace || isFalse;
             };
-            Func<PyObject, PyTuple> selector = x =>
+            Func<DateTime, PyTuple> selector = x =>
             {
-                list[list.Count - 1] = x;
+                list[list.Count - 1] = x.ToPython();
                 return new PyTuple(list.ToArray());
             };
-
+            // creating the pandas MultiIndex is expensive so we keep a cash
+            var indexCache = new Dictionary<List<DateTime>, dynamic>(new ListComparer<DateTime>());
             using (Py.GIL())
             {
                 // Returns a dictionary keyed by column name where values are pandas.Series objects
@@ -299,8 +280,13 @@ namespace QuantConnect.Python
                     var values = kvp.Value.Item2;
                     if (values.All(filter)) continue;
 
-                    var tuples = kvp.Value.Item1.Select(selector).ToArray();
-                    var index = _pandas.MultiIndex.from_tuples(tuples, names: splitNames);
+                    dynamic index;
+                    if (!indexCache.TryGetValue(kvp.Value.Item1, out index))
+                    {
+                        var tuples = kvp.Value.Item1.Select(selector).ToArray();
+                        index = _pandas.MultiIndex.from_tuples(tuples, names: splitNames);
+                        indexCache[kvp.Value.Item1] = index;
+                    }
 
                     pyDict.SetItem(kvp.Key, _pandas.Series(values, index));
                 }
@@ -315,11 +301,11 @@ namespace QuantConnect.Python
         /// <param name="key">The key of the value to get</param>
         /// <param name="time"><see cref="DateTime"/> object to add to the value associated with the specific key</param>
         /// <param name="input"><see cref="Object"/> to add to the value associated with the specific key</param>
-        private void AddToSeries(string key, PyObject time, object input)
+        private void AddToSeries(string key, DateTime time, object input)
         {
             if (input == null) return;
 
-            Tuple<List<PyObject>, List<object>> value;
+            Tuple<List<DateTime>, List<object>> value;
             if (_series.TryGetValue(key, out value))
             {
                 value.Item1.Add(time);